Output grid spacing as double, not real, values
This correctly matches the BMI specification.

+++ bmi/bmi.f90
@@ -218,7 +218,7 @@ module bmif
        import :: bmi
        class (bmi), intent(in) :: self
        integer, intent(in) :: grid_id
-       real, dimension(:), intent(out) :: grid_spacing
+       double precision, dimension(:), intent(out) :: grid_spacing
        integer :: bmi_status
      end function bmif_get_grid_spacing
